Bugfix: Type error for modifier

+++ include/db/dbcard.h
@@ -40,7 +40,7 @@ typedef struct PlayedCard {
 int GetCardId(char* card_title);
 char* GetCardPath(unsigned int card_id);
 int GetCardPriceModifier(unsigned int card_id);
-float GetCardModifierLength(unsigned int card_id);
+unsigned int GetCardModifierLength(unsigned int card_id);
 char* GetCardDescription(unsigned int card_id);
 int GetNumOfPlayerPositiveCards();
 int GetNumOfPlayerNegativeCards();
